+++ src/dynamic-form.index.ts
@@ -1,0 +1,13 @@
+import { ODynamicFormComponent } from './o-dynamic-form.component';
+import { ODFComponentComponent } from './o-dynamic-form-component.component';
+import { ODFElementComponent } from './o-dynamic-form-element.component';
+import { ODFElementOptionsComponent } from './o-dynamic-form-element-options.component';
+import { ODropZoneComponent } from './dropzone/o-drop-zone.component';
+
+export const O_DYNAMICFORM_COMPONENTS: any = [
+  ODynamicFormComponent,
+  ODFComponentComponent,
+  ODFElementComponent,
+  ODFElementOptionsComponent,
+  ODropZoneComponent
+];