+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 
 ### PEER-DEPENDENCY UPDATES ###
-* **Updated**:   ontimize-web-ng2@1.2.2
+* **Updated**:   ontimize-web-ngx@1.2.2
 
 ### BREAKING CHANGES
 
@@ -19,13 +19,13 @@
 * **DynamicFormModule**: exporting '*ODynamicFormComponent*'.
 
 ### PEER-DEPENDENCY UPDATES ###
-* **Updated**:   ontimize-web-ng2@1.2.1
+* **Updated**:   ontimize-web-ngx@1.2.1
 
 ### BREAKING CHANGES
 
 ### Bug Fixes
 * **Mode bugs**: disabling drag and drop when '*edit-mode*' is false.
-* **o-form**: '*form-definition*' input bugs fixed ([#1](https://github.com/OntimizeWeb/ontimize-web-ng2/issues/1)) ([18865eb](https://github.com/OntimizeWeb/ontimize-web-ng2/commit/18865eb))
+* **o-form**: '*form-definition*' input bugs fixed ([#1](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/1)) ([18865eb](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/18865eb))